Fix compilation

+++ src/lwjgl/java/net/minestom/demo/largeframebuffers/Demo.java
@@ -68,7 +68,7 @@ public class Demo {
     private static void createFrame(Instance instance, int id, int x, int y, int z) {
         EntityItemFrame itemFrame = new EntityItemFrame(new Position(x, y, z), EntityItemFrame.ItemFrameOrientation.NORTH);
         itemFrame.getPosition().setYaw(180f);
-        ItemStack map = new ItemStack(Material.FILLED_MAP, (byte) 1, 1);
+        ItemStack map = new ItemStack(Material.FILLED_MAP, (byte) 1, (short) 1);
         map.setItemMeta(new MapMeta());
         itemFrame.setItemStack(map);
         itemFrame.setInstance(instance);
